requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
openai>=1.0
python-dotenv
